% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_head.r
\name{dataset_head}
\alias{dataset_head}
\title{Get the head of the dataset in the weles}
\usage{
dataset_head(dataset_id, n = 5)
}
\arguments{
\item{dataset_id}{the dataset hash}

\item{n}{number of rows to show}
}
\value{
top n rows of the dataset
}
\description{
This tool allows you to view the head of the dataset in the weles.
}
\examples{
library("weles")

dataset_head('aaaaaaaaaaaaaaaaaaaaaaaaaaaaa', 10)

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
