% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_info.r
\name{dataset_info}
\alias{dataset_info}
\title{Get an info about the dataset in weles}
\usage{
dataset_info(dataset_id)
}
\arguments{
\item{model_name}{Name of the model in weles, character}
}
\value{
named list containing all meta data about model
}
\description{
This tool is used for getting a meta data about the model that is already uploaded to weles.
}
\examples{
library("weles")

dataset_info("aaaaaaaaaaaaaaaaaaaaaaa")
dataset_info("aaaaaaaaaaaaaaaaaaaaaaa")$number_of_rows
dataset_info("aaaaaaaaaaaaaaaaaaaaaaa")$number_of_columns
dataset_info("aaaaaaaaaaaaaaaaaaaaaaa")$columns

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
