% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_upload.r
\name{dataset_upload}
\alias{dataset_upload}
\title{Upload the dataset to the weles}
\usage{
dataset_upload(data, data_name, data_desc)
}
\arguments{
\item{data}{the data frame to upload or path}

\item{data_name}{name of the dataset that will be visible in the weles}

\item{data_desc}{description of the dataset}
}
\value{
information if uploading the data was successful
}
\description{
This function uploads the dataset and needed metadata to the weles.
}
\examples{
library("weles")

dataset_upload(X, 'data name', 'Example user', 'example password')

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
