% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_get.r
\name{datasets_get}
\alias{datasets_get}
\title{Get the dataset from weles}
\usage{
datasets_get(dataset_id)
}
\arguments{
\item{dataset_id}{hash of the dataset}
}
\value{
data frame
}
\description{
You can use this function to download the dataset from weles as a data frame
}
\examples{
library("weles")

datasets_get('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
