% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.r
\name{model_info}
\alias{model_info}
\title{Get an info about the model in weles}
\usage{
model_info(model_name)
}
\arguments{
\item{model_name}{Name of the model in weles, character}
}
\value{
named list containing all meta data about model
}
\description{
This tool is used for getting a meta data about the model that is already uploaded to weles.
}
\examples{
library("weles")

model_info("example_model")
model_info("example_model")$model
model_info("example_model")$data
model_info("example_model")$data$dataset_id

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
