% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_status.r
\name{model_status}
\alias{model_status}
\title{Get the status of the uploading model}
\usage{
model_status(task_id, interactive = TRUE)
}
\arguments{
\item{task_id}{id of the uploading, it is always returned by function models_upload}

\item{interactive}{display the progressbar}
}
\value{
named list containing meta data about uploading
}
\description{
If you know the uploading id, you can check the progress.
}
\examples{
library("weles")

model_status("aaaaaaaaaaaaaaaaaaaaaaaaa")
model_status("aaaaaaaaaaaaaaaaaaaaaaaaa")$model_existed
model_status("aaaaaaaaaaaaaaaaaaaaaaaaa", interactive = FALSE)

}
