% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_audit.r
\name{models_audit}
\alias{models_audit}
\title{Make an audit of the model in the weles}
\usage{
models_audit(model_name, measure, data, target, data_name = NA,
  data_desc = NA)
}
\arguments{
\item{model_name}{name of the model in the weles, character}

\item{measure}{name of the measure used in the audit, character}

\item{data}{data frame to make an audit on or path or hash of already uploaded dataset}

\item{target}{name of the target column in the dataset}

\item{data_name}{name of the dataset that will be visible in the weles, unnecessary if data is a hash}

\item{data_desc}{description of the dataset, unnecessary if data is a hash}
}
\value{
result of the audit or information if somethin went wrong
}
\description{
You can use this function to audit in different ways models already uploaded in the weles.
}
\examples{
\\code{
library("weles")

models_audit('example_model', 'mae', 'Example user', 'example password', iris, 'Species', 'iris', 'Flowers')
models_audit('example_model', 'acc', 'Example user', 'example password', 'aaaaaaaaaaaaaaaaaaaaaaaaaa', 'Species')
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
