% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_predict.r
\name{models_predict}
\alias{models_predict}
\title{Make a prediction using weles model}
\usage{
models_predict(model_name, X, pred_type = "exact",
  prepare_columns = TRUE)
}
\arguments{
\item{model_name}{name of the model in weles}

\item{X}{data to make a prediction of, must have named columns, may be path to *.csv* file (must contatin **/** sign) or *hash* of already uploaded data,
if X is an object and prepare_columns is True, columns' names will be fetched automatically}

\item{pred_type}{type of prediction, 'exact' or 'prob'}

\item{prepare_columns}{if X is an object then columns' names will be fetched automatically}
}
\description{
This tool allows you to make a prediction with model in weles.
}
\examples{
\\code {
library("weles")

models_predict("example_model", iris[,-5])
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
