% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_requirements.r
\name{models_requirements}
\alias{models_requirements}
\title{Get package requirements for the model in weles}
\usage{
models_requirements(model)
}
\arguments{
\item{model}{name of the model in weles, character}
}
\value{
named list with requirements
}
\description{
You can use this function to download requirements of the model
}
\examples{
library("weles")

models_requirements("example_model")
models_requirements("example_model")$pandas

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
