% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_search.r
\name{models_search}
\alias{models_search}
\title{Find model that interests you the most}
\usage{
models_search(language = NA, language_version = NA, row = NA,
  column = NA, missing = NA, classes = NA, owner = NA,
  tags = c(), regex = NA)
}
\arguments{
\item{language}{search only among models written in this language}

\item{language_version}{what language version should be model written, '<n;' '>n;' '=n;' '>a;<b;'}

\item{row}{parameter descibing number of rows in training dataset, '<n;' '>n;' '=n;' '>a;<b'}

\item{column}{parameter descibing number of columns in training dataset, '<n;' '>n;' '=n;' '>a;<b'}

\item{missing}{parameter descibing number of missing values in training dataset, '<n;' '>n;' '=n;' '>a;<b'}

\item{classes}{parameter descibing number of classes in training dataset, '<n;' '>n;' '=n;' '>a;<b'}

\item{owner}{show only models created by this user}

\item{tags}{vector of tags, should be all strings}

\item{regex}{regex for searching names of models}
}
\value{
vector of models' names satisfying those restrictions
}
\description{
Function allows you advanced search of models in weles. If all parameters are default then returns all models' name in weles.
}
\examples{
\\code{
library("weles")

models_search(tags = c('example', 'easy'))

models_search(row='<15000;', tags = c('example', 'easy'))

models_search(column='>10;<15;', owner='Example user')

models_search(language='python', language_version='3.6.8', row='>1000;<10000;', column='=14;', classes='=2;', missing='=0;', owner='Example user', tags=
c('example', 'easy'), regex='^R')
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
