% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_upload.r
\name{models_upload}
\alias{models_upload}
\title{Upload the model to weles}
\usage{
models_upload(model, model_name, model_desc, target, tags, train_dataset,
  train_dataset_name = NA, dataset_desc = NA, requirements_file = NA)
}
\arguments{
\item{model}{mlr, caret or parsnip model, or path to it written as '.RDS' file}

\item{model_name}{name of the model that will be visible in the weles}

\item{model_desc}{description of the model}

\item{target}{name of the target column}

\item{tags}{vector of model's tags, should be vector of strings}

\item{train_dataset}{training dataset with named columns and target column as the last one, or path to *.csv* file (must contain **/** sign) or hash of already
uploaded data}

\item{train_dataset_name}{name of the training dataset that will be visible in the database}

\item{dataset_desc}{description of the dataset}
}
\description{
This tool allows you to upload your mlr, caret or parsnip model to model governance base weles.
}
\examples{
\\code{
library("weles")
library("mlr")

task = makeClassifTask(data = iris, target = 'Species')
model = makeLearner("classif.randomForest")
model = train(model, task)

models_upload(model, "example_model", "This is an example model", 'Species', c('example', 'easy'), iris, "example_training_data", 'This is an example data', 'Example user', 'example password')
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
