% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_create.r
\name{user_create}
\alias{user_create}
\title{Create user in the weles}
\usage{
user_create()
}
\value{
Information if creating an account was successful.
}
\description{
This function is used to create an user account. You need this to have an access to all features in weles.
}
\examples{
\\code{
library("weles")

user_create("Example user", "example password", "example_mail@gmail.com")
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
