% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_status.r
\name{models_status}
\alias{models_status}
\title{Get an info about the model in weles}
\usage{
models_status(task_id, interactive = TRUE)
}
\arguments{
\item{model_name}{Name of the model in weles, character}
}
\value{
named list containing all meta data about model
}
\description{
This tool is used for getting a meta data about the model that is already uploaded to weles.
}
\examples{
library("weles")

models_info("example_model")
models_info("example_model")$model
models_info("example_model")$data
models_info("example_model")$data$dataset_id

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
